Refactor chat endpoint to use ChatService
